adjust filenames and seperate voting functionality into it's own component
